requirements.txt
alembic==1.3.1
azure-cognitiveservices-speech==1.8.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.17.4
python-dateutil==2.8.1
python-editor==1.0.4
scipy==1.3.3
six==1.13.0
sounddevice==0.3.14
SQLAlchemy==1.3.11
Werkzeug==0.16.0
WTForms==2.2.1
